+++ deploy-clojars.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+op run --env-file="clojars.env" -- clojure -T:build deploy