+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file. This change
 
 [Unreleased]: https://github.com/net.curiousprogrammer/email-check/compare/0.1.0...HEAD
 
+## 0.1.1 - 2023-02-07
+
+- more robust email domain check
+- return `:format-error` describing the syntactical issues as per `validate`.
+
 ## 0.1.0 - 2023-02-07
 
 ### Added